Clicking on a template tag e.g. [[field_name]] now adds the template tag to the HTMLArea editor in the template views. This is not currently working in source code mode of the editor.
Added equivalent of use_html_editor() and print_editor_config() from
/lib/weblib.php to textarea field class. We use a copy in that class instead of
calling the functions in /lib/weblib.php because we want the data module to be
compatible with Moodle 1.5. This is not ideal, but it is a necessary evil.

Updated the textarea field class to use the correct path to
/lib/editor/htmlarea/htmlarea.php.