MDL-42023 assign: Add Edit PDF plugin to core.
This plugin is heavily based on the uploadpdf plugin from
Davo Smith <git@davosmith.co.uk>. This is a fantastic plugin
that provides a great benefit for teachers marking assignments.

A big thankyou to Davo for writing the plugin and helping us to
get it integrated to core.

The plugin provides a pdf editing interface in the teachers browser
that lets them add comments and annotations to a students assignment
submission. The comment and annotations are then added to a pdf version
of the students submission which is send back to the student so they can
download it, or read it online.

This plugin has been primarily worked on by Damyon Wiese, Jerome Mouneyrac
and Barbara Ramiro. We have also had great feedback and suggestions from others
including the entire frontend team, Martin and Frederic Massart.