MDL-42402 EditPDF: CSS Improvements for editing toolbar.
The main change here is to prevent the default rendering of a select list for the page
selector, so that it looks more consistent with the rest of the toolbar on different
browsers/os combinations.

This is done with some vendor specific css rules targeting webkit, mozilla and ie.

There are also some minor tweaks to vertically align the images in the buttons and reduce the padding
above and below the toolbar.