MDL-42351 EditPDF: Fix jiggling of relatively positioned elements
When calling setX/setY with YUI - YUI takes into account the position style
of the element. If none is set it assumes position relative. Position relative
is bad because relatively positioned elements still take up space in the page,
and all the elements after that node are positioned, taking into account the size
of that node. So when a node is resized or deleted, all nodes after that one in
the list of drawables are affected.

The magic fix - is just to set position: absolute on the drawables before calling setX/setY.